backend/requirements.txt

fastapi==0.112.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.2
numpy>=1.22.4,<2.0.0
akshare==1.16.72
tushare==1.2.89
python-dateutil==2.9.0.post0
pydantic==2.8.2
requests==2.32.3
apscheduler==3.10.4
statsmodels==0.14.2
scikit-learn==1.5.1
scipy>=1.10.0,<1.14.0
python-dotenv==1.1.1

# News processing dependencies
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pymongo==4.10.1

# Redis support
redis==5.0.8
